backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemygit